pyproject.toml
[project]
name = "uber-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=38.2.0",
    "fastapi>=0.128.8",
    "flask-cors>=6.0.1",
    "geohash2>=1.1",
    "httpx>=0.28.1",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "osmnx>=2.0.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]
